makeDist

turn vectors of a values and their distribution function values into an ecdf. Vectors should be the same length and both increasing.

makeDist(x, Fx, sorted = FALSE, rearrange = FALSE)

Arguments

x
vector of values

Fx
vector of the distribution function values

sorted
boolean indicating whether or not x is already sorted; computation is somewhat faster if already sorted

rearrange
boolean indicating whether or not should monotize distribution function

Value

ecdf

Examples

y <- rnorm(100)
y <- y[order(y)]
u <- runif(100)
u <- u[order(u)]
F <- makeDist(y,u)